+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changes
 
 ## UNRELEASED
 
+* Feature: Async JavaScript configuration function.
 * Dependencies: Various updates.
 * Test: Upgrade `mock-fs` and re-enable Node16+
   [#39](https://github.com/FormidableLabs/trace-pkg/issues/39)